+++ oceanbase/src/common/ob_probability_random.cpp
@@ -57,7 +57,7 @@ int32_t ObProbabilityRandom::random(void) const
   int32_t ret = -1;
   if (true == init_)
   {
-    int32_t random_percent = ::random();
+    int32_t random_percent = static_cast<int32_t>(::random());
     tbsys::CRLockGuard lock(lock_);
     ret = index_[random_percent % index_count_];
   }
@@ -100,7 +100,7 @@ int32_t ObStalessProbabilityRandom::random(const int32_t percent[], const int32_
   if ((count > 0) && (sum > 0))
   {
     int32_t step_percent = 0;
-    int32_t random_percent = ::random() % sum;
+    int32_t random_percent = static_cast<int32_t>(::random() % sum);
     for (int32_t i = 0; i < count; ++i)
     {
       step_percent += percent[i];
